rename channels module to device module
