+++ qiskit/pulse/commands/sample_pulse.py
@@ -10,7 +10,7 @@ Sample pulse.
 """
 import numpy as np
 
-from qiskit.pulse.channels import OutputChannel
+from qiskit.pulse.device import OutputChannel
 from qiskit.pulse.timeslots import Interval, Timeslot, TimeslotCollection
 from qiskit.pulse.exceptions import PulseError
 from .instruction import Instruction
