+++ qiskit/pulse/device/device_specification.py
