+++ qiskit/pulse/device/output_channel.py
