+++ qiskit/pulse/device/pulse_channel.py
