+++ qiskit/pulse/device/qubit.py
